Fix issue with audio studdering
I'm not sure if this actually fixed it or not. Seems like setting
these environmental variables should not make a difference.

Another possibility is that restarting the VM after changing the
sound sample rates might have done something to stop the studdering.

I'm pretty sure I had already tried rebooting but now the audio is
crystal clear. Hopefully it remains that way.